melhorias

+++ admin/editExercicios.php
@@ -2,7 +2,7 @@
 require_once '../vendor/autoload.php';
 
 $con = new \App\Model\Conexao('exercicios');
-
+$alertaArquivo  = '';
 // buscando informações para preencher os inputs
 if (isset($_GET['id']) and !empty($_GET['id']) and is_numeric($_GET['id'])) {
 
@@ -12,35 +12,42 @@ if (isset($_GET['id']) and !empty($_GET['id']) and is_numeric($_GET['id'])) {
 if (isset($_POST['btnSalvar']) && !empty($_POST['btnSalvar'])) {
 
     $anexo = [];
+    $dados = filter_input_array(INPUT_POST, FILTER_DEFAULT);
 
     $formatosPermitidos = array("png", "jpge", "jpg", "svg", "mp4", "mkv", "avi", "wmv", "wma", "rmvb", "mov", "3gp", "flv");
     $extensao = PATHINFO($_FILES['arquivo']['name'], PATHINFO_EXTENSION);    // EXTENSÃO DO ARQUIVO
 
-    if (in_array($extensao, $formatosPermitidos)) {
+    if (!empty($_FILES['arquivo']['tmp_name'])) {
+        if (in_array($extensao, $formatosPermitidos)) {
 
-        $pasta = "arquivos/exercicios/";
-        $caminhoTemporario = $_FILES['arquivo']['tmp_name'];
-        $novoNome = uniqid() . ".$extensao";
+            $pasta = "arquivos/exercicios/";
+            $caminhoTemporario = $_FILES['arquivo']['tmp_name'];
+            $novoNome = uniqid() . ".$extensao";
 
-        // UPLOAD
-        move_uploaded_file($caminhoTemporario, $pasta . $novoNome);
-    }
-
-    $dados = filter_input_array(INPUT_POST, FILTER_DEFAULT);
-    $arquivo = $con->read('id= ' . $dados['id']);
-    $pasta = "arquivos/exercicios/";
-
-    // CASO QUEIRA ATUALIZAR O ARQUIVO
-    if (!empty($novoNome)) {
-        foreach ($arquivo as $v) {
-            unlink($pasta . "/" . $v['anexo']);
+            // UPLOAD
+            move_uploaded_file($caminhoTemporario, $pasta . $novoNome);
+        } else {
+            $infoInputs = $con->read('id= ' . $dados['id']);
+            $alertaArquivo  = 'Arquivo Inválido! Escolha um arquivo com formato permitido.';
         }
-        $anexo['anexo'] = $novoNome;
     }
-    $atualizacao = $con->update('id= ' . $dados['id'], [
-        'nome'    => $dados['nome'],
-        'descricao' => $dados['descricao']
-    ] + $anexo);
+
+    if (empty($alertaArquivo)) {
+        $arquivo = $con->read('id= ' . $dados['id']);
+        $pasta = "arquivos/exercicios/";
+
+        // CASO QUEIRA ATUALIZAR O ARQUIVO
+        if (!empty($novoNome)) {
+            foreach ($arquivo as $v) {
+                unlink($pasta . "/" . $v['anexo']);
+            }
+            $anexo['anexo'] = $novoNome;
+        }
+        $atualizacao = $con->update('id= ' . $dados['id'], [
+            'nome'    => $dados['nome'],
+            'descricao' => $dados['descricao']
+        ] + $anexo);
+    }
 }
